[MOV] mail: move '_mail_track' to tracking mixin
'_mail_track' is a method that generates tracking values based on
changes. It is located on BaseModel just to have it available on
models doing some tracking but not inheriting from mail;tnead
(see odoo/odoo#56631).

We move it into 'mail.track' mixin to make it coherent with new
model. This should closes issues like model not inheriting from
mail.thread having partially access to tracking methods in some
hackish ways.

Now generating value should be done only on models inheriting at least
from 'mail.track.mixin'.

Task-5935695 ([mail, various] Cleanup and test tracking usage)
Prepares Task-3645865 ([mail] In-body tracking)
